refactor(quality): modularize quality management components and update configurations
