+++ config/detekt/detekt-baseline.xml
@@ -3,6 +3,7 @@
   <ManuallySuppressedIssues/>
   <CurrentIssues>
     <ID>LongMethod:BomManager.kt$BomManager$private fun DependencyHandler.configureCommonBoms()</ID>
+    <ID>LongMethod:SpotlessConfigurator.kt$internal fun Project.configureSpotless( quality: QualityExtension, libs: VersionCatalog, )</ID>
     <ID>TooManyFunctions:PluginExtension.kt$PluginExtension</ID>
   </CurrentIssues>
 </SmellBaseline>
