+++ convention-plugin/src/main/kotlin/io/github/aaravmahajanofficial/internal/quality/DetektConfigurator.kt
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2025 Aarav Mahajan
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+package io.github.aaravmahajanofficial.internal.quality
+
+import io.github.aaravmahajanofficial.extensions.quality.QualityExtension
+import io.github.aaravmahajanofficial.utils.hasKotlinSources
+import io.github.aaravmahajanofficial.utils.resolveConfigFile
+import io.github.aaravmahajanofficial.utils.versionFromCatalogOrFail
+import io.gitlab.arturbosch.detekt.Detekt
+import io.gitlab.arturbosch.detekt.DetektPlugin
+import io.gitlab.arturbosch.detekt.extensions.DetektExtension
+import org.gradle.api.Project
+import org.gradle.api.artifacts.VersionCatalog
+import org.gradle.kotlin.dsl.configure
+import org.gradle.kotlin.dsl.withType
+
+internal fun Project.configureDetekt(
+    quality: QualityExtension,
+    libs: VersionCatalog,
+) {
+    if (!quality.detekt.enabled.get() || !hasKotlinSources()) return
+
+    pluginManager.apply(DetektPlugin::class.java)
+
+    configure<DetektExtension> {
+        toolVersion = versionFromCatalogOrFail(libs, "detekt")
+        autoCorrect = quality.detekt.autoCorrect.get()
+        buildUponDefaultConfig = true
+        isIgnoreFailures = !quality.detekt.failOnViolation.get()
+        source.setFrom(files("src/main/kotlin", "src/test/kotlin").plus(quality.detekt.source.get()))
+        config.setFrom(resolveConfigFile("detekt", "detekt.yml"))
+        baseline = resolveConfigFile("detekt", "detekt-baseline.xml").asFile
+        parallel = true
+    }
+    tasks.withType<Detekt>().configureEach { detekt ->
+        detekt.reports {
+            it.xml.required.set(true)
+            it.html.required.set(true)
+            it.sarif.required.set(true)
+        }
+    }
+    tasks.named("check").configure { t ->
+        t.dependsOn("detekt")
+    }
+}
