+++ convention-plugin/src/main/kotlin/io/github/aaravmahajanofficial/internal/quality/DokkaConfigurator.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2025 Aarav Mahajan
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+package io.github.aaravmahajanofficial.internal.quality
+
+import io.github.aaravmahajanofficial.extensions.quality.QualityExtension
+import io.github.aaravmahajanofficial.utils.hasKotlinSources
+import org.gradle.api.Project
+import org.jetbrains.dokka.gradle.DokkaPlugin
+
+internal fun Project.configureDokka(quality: QualityExtension) {
+    if (!quality.dokka.enabled.get() || !hasKotlinSources()) return
+
+    pluginManager.apply(DokkaPlugin::class.java)
+}
